requirements.txt
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0 
python-jobspy>=1.1.77
pathlib>=1.0.1 